Fix arkolia studio imports after module relocation
