+++ apps/web/js/views/studio/solidity/solidity-arkolia.js
@@ -1,16 +1,16 @@
-import { searchFrenchCommunes, fetchFrenchAltitude } from "../../services/georisques-service.js";
-import { getCantonByCommuneCode } from "../../services/zoning/canton-service.js";
-import { getWindRegionsByDepartmentCode } from "../../services/zoning/wind-regions-service.js";
-import { getSnowRegionsByDepartmentCode } from "../../services/zoning/snow-regions-service.js";
-import { getWindZoneByDepartmentAndCanton } from "../../services/zoning/wind-canton-regions-service.js";
-import { getSnowZoneByDepartmentAndCanton } from "../../services/zoning/snow-canton-regions-service.js";
-import { getFrostDepthByDepartmentCode } from "../../services/zoning/frost-depth-service.js";
-import { escapeHtml } from "../../utils/escape-html.js";
-import { buildGoogleMapsPlaceEmbedUrl, hasGoogleMapsEmbedApiKey } from "../../services/google-maps-embed-service.js";
-import { registerProjectPrimaryScrollSource } from "../project-shell-chrome.js";
-import { svgIcon } from "../../ui/icons.js";
-import { store } from "../../store.js";
-import { renderGhActionButton } from "../ui/gh-split-button.js";
+import { searchFrenchCommunes, fetchFrenchAltitude } from "../../../services/georisques-service.js";
+import { getCantonByCommuneCode } from "../../../services/zoning/canton-service.js";
+import { getWindRegionsByDepartmentCode } from "../../../services/zoning/wind-regions-service.js";
+import { getSnowRegionsByDepartmentCode } from "../../../services/zoning/snow-regions-service.js";
+import { getWindZoneByDepartmentAndCanton } from "../../../services/zoning/wind-canton-regions-service.js";
+import { getSnowZoneByDepartmentAndCanton } from "../../../services/zoning/snow-canton-regions-service.js";
+import { getFrostDepthByDepartmentCode } from "../../../services/zoning/frost-depth-service.js";
+import { escapeHtml } from "../../../utils/escape-html.js";
+import { buildGoogleMapsPlaceEmbedUrl, hasGoogleMapsEmbedApiKey } from "../../../services/google-maps-embed-service.js";
+import { registerProjectPrimaryScrollSource } from "../../project-shell-chrome.js";
+import { svgIcon } from "../../../ui/icons.js";
+import { store } from "../../../store.js";
+import { renderGhActionButton } from "../../ui/gh-split-button.js";
 
 const DEFAULT_IDENTITY = {
   length: "",
